* news carousel adjustments

* adjust explore and tag pages

* work header fixes

* profile header updates

* adjust profile home page

* sizes down work stats on smaller displays

* adjusting works and blogs lists

* Update +layout.svelte

* adding optimizations to section list

* Update +page.svelte

* Update +page.svelte

* Feat/email and recovery (#136)

* feat: password resets and email confirmations

* updating controllers with confirmation guard

* API routes for features now in place

* password resets are in

* email confirmations are now in

* update readme

+++ Sources/App/Controllers/Accounts/AccountController.swift
@@ -7,6 +7,7 @@ import Vapor
 struct AccountController: RouteCollection {
     func boot(routes: RoutesBuilder) throws {
         let accounts = routes.grouped(IdentityGuard(needs: [.user])).grouped("accounts")
+        let accountsNoAuth = routes.grouped("accounts")
 
         accounts.get("fetch-profile", ":profileId") { request async throws -> Profile in
             let profileId = request.parameters.get("profileId")!
@@ -17,23 +18,36 @@ struct AccountController: RouteCollection {
             try await request.accountService.fetchProfiles()
         }
 
-        accounts.grouped(BannedGuard()).post("create-profile") { request async throws -> Profile in
+        accounts.grouped([ConfirmationGuard(), BannedGuard()]).post("create-profile") { request async throws -> Profile in
             try Profile.ProfileForm.validate(content: request)
             let profileForm = try request.content.decode(Profile.ProfileForm.self)
             return try await request.accountService.createProfile(with: profileForm)
         }
 
-        accounts.grouped(BannedGuard()).delete("delete-profile", ":profileId") { request async throws -> Response in
+        accounts.grouped([ConfirmationGuard(), BannedGuard()]).delete("delete-profile", ":profileId") { request async throws -> Response in
             let profileId = request.parameters.get("profileId")!
             try await request.accountService.deleteProfile(profileId)
             return Response(status: .ok)
         }
         
-        accounts.grouped(BannedGuard()).post("file-report") { request async throws -> Response in
+        accounts.grouped([ConfirmationGuard(), BannedGuard()]).post("file-report") { request async throws -> Response in
             try AccountReport.ReportForm.validate(content: request)
             let formInfo = try request.content.decode(AccountReport.ReportForm.self)
-            print(formInfo)
             return try await request.adminService.reportUser(with: formInfo)
+        }
+        
+        accounts.grouped(BannedGuard()).get("send-confirmation-email") { request async throws -> Response in
+            try await request.accountService.sendConfirmationEmail()
+        }
+        
+        accountsNoAuth.patch("confirm-email") { request async throws -> Response in
+            try EmailConfirmation.EmailConfirmationForm.validate(content: request)
+            let formInfo = try request.content.decode(EmailConfirmation.EmailConfirmationForm.self)
+            return try await request.accountService.confirmEmail(with: formInfo)
+        }
+        
+        accounts.grouped(BannedGuard()).patch("agree-to-terms") { request async throws -> ClientAccount in
+            try await request.accountService.agreeToTerms()
         }
     }
 }
